(DOCSP-26992) Add Realm React support to 'Relationships'
