+++ examples/react-native/package.json
@@ -21,8 +21,9 @@
     "@babel/preset-typescript": "^7.18.6",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/jest-native": "^4.0.13",
+    "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/react-native": "^11.2.0",
     "@tsconfig/react-native": "^2.0.2",
     "@types/jest": "^26.0.23",
