+++ source/sdk/react-native/realm-database/crud/read.txt
@@ -123,7 +123,7 @@ In the following example of a ``TaskList`` component, we:
 
    To filter a query based on a property of an :ref:`embedded object
    <react-native-embedded-objects>` or a :ref:`related object
-   <react-native-client-relationships>`, use dot-notation as if it were in a regular,
+   <react-native-relationships>`, use dot-notation as if it were in a regular,
    nested object.
 
 .. seealso::
@@ -178,5 +178,5 @@ Finally, we map through each list of tasks and render them in the UI.
    
    To sort a query based on a property of an :ref:`embedded object
    <react-native-embedded-objects>` or a :ref:`related object
-   <react-native-client-relationships>`, use dot-notation as if it were in a
+   <react-native-relationships>`, use dot-notation as if it were in a
    regular, nested object.
