+++ source/sdk/react-native/realm-database/crud/update.txt
@@ -66,7 +66,7 @@ In the following example of a ``TaskItem`` component, we:
 .. tip:: Update Related and Embedded Objects
    
    To update a property of an :ref:`embedded object <react-native-embedded-objects>` or
-   a :ref:`related object <react-native-client-relationships>`, modify the property with
+   a :ref:`related object <react-native-relationships>`, modify the property with
    :mdn:`dot-notation or bracket-notation
    <Web/JavaScript/Reference/Operators/Property_accessors>` as if it were in a
    regular, nested object.
