+++ source/sdk/react-native/realm-database/define-a-realm-object-model.txt
@@ -20,7 +20,7 @@ Key Concept: Object Types & Schemas
 -----------------------------------
 
 Every Realm object conforms to a specific **object type**, which is essentially
-a class that defines the properties and :ref:`relationships <react-native-client-relationships>`
+a class that defines the properties and :ref:`relationships <react-native-relationships>`
 for objects of that type using a pre-defined schema. Realm guarantees that all
 objects in a realm conform to the schema for their object type and validates
 objects whenever they're created, modified, or deleted.
